fixed eports in actions file
