+++ src/app/deals/add/page.tsx
@@ -1,35 +1,10 @@
 import DealForm from './DealForm';
-import { Metadata } from 'next';
-import { getXataClient } from '@/xata';
-import { redirect } from 'next/navigation';
-import { z } from 'zod';
 
 //TODO: why doesn't this work with server component page?
 // export const metadata: Metadata = {
 //   title: 'Add a Black Friday Deal',
 //   description: 'Share the best deals that you know developers will love!',
 // };
-
-export enum Category {
-  Ebook = 'ebook',
-  Video = 'video',
-  Tool = 'tool',
-  Conference = 'conference',
-  Misc = 'misc',
-}
-
-export const dealSchema = z.object({
-  name: z.string(),
-  link: z.string().url(),
-  description: z.string(),
-  startDate: z.coerce.date(),
-  endDate: z.coerce.date(),
-  coupon: z.string().optional(),
-  couponPercentage: z.number().optional(),
-  email: z.string().email().optional(),
-  //TODO: don't replicate array
-  category: z.enum(['misc', 'ebook', 'video', 'tool', 'conference']),
-});
 
 export default function AddDealPage() {
   return (
